Remove switch view
