--- src/app/Badge/style.scss
@@ -1,10 +1,0 @@
-.badge {
-  padding: 6px;
-  position: relative;
-  top: -3px;
-}
-
-.badge__icon {
-  position: relative;
-  top: -2px;
-}
