--- src/app/SearchHit/style.scss
@@ -1,31 +1,0 @@
-$card-size: 12rem;
-$card-size-large: 10rem;
-$card-image-size: 80px;
-$card-image-size-large: 96px;
-
-.search-hit {
-  height: $card-size;
-  border: 1px solid transparent;
-
-  @include break(large) {
-    height: $card-size-large;
-  }
-
-  @include hover-transition(border-color);
-
-  &:hover {
-    border-color: material-color('blue', '500');
-  }
-}
-
-.search-hit__image {
-  width: $card-image-size;
-  height: $card-image-size;
-
-  @include break(large) {
-    width: $card-image-size-large;
-    height: $card-image-size-large;
-  }
-
-  object-fit: contain;
-}
