+++ src/app/SearchHits/index.js
@@ -1,15 +1,11 @@
 import InfiniteScroll from 'react-infinite-scroll-component'
 import React, {createElement} from 'react'
-import classnames from 'classnames'
 
 import SearchStats from '../SearchStats'
 import ProductCard from '../ProductCard'
-import SearchHit from '../SearchHit'
 
 export default (props) => {
-  const {get, hits, refine, hasMore} = props
-  const isGrid = get('hitComponent') === 'grid'
-  const childComponent = isGrid ? ProductCard : SearchHit
+  const {hits, refine, hasMore} = props
 
   return (
     <InfiniteScroll
@@ -19,11 +15,9 @@ export default (props) => {
         >
       <div data-app='search-hits'>
         <SearchStats {...props} />
-        <div className={classnames('overflow-hidden', {
-          'flex justify-center flex-wrap': isGrid
-        })}>
+        <div className='overflow-hidden flex justify-center flex-wrap'>
           {hits.map((product, key) => (
-            createElement(childComponent, {product, key, ...props})
+            createElement(ProductCard, {product, key, ...props})
           ))}
         </div>
       </div>
