--- src/app/SwitchView/style.scss
@@ -1,11 +1,0 @@
-$border-radius: .25rem;
-
-.switch-view__list {
-  border-top-left-radius: $border-radius;
-  border-bottom-left-radius: $border-radius;
-}
-
-.switch-view__grid {
-  border-top-right-radius: $border-radius;
-  border-bottom-right-radius: $border-radius;
-}
